==> 1-Data-Loading.ipynb <==
# Obtain data

In this Notebook, we retrieve the data and save it as a JSON file. 

Whilst the data was said to be in JSON format, when running it through a JSON linter, we noticed that it was not proper JSON due to the use of single quotes instead of double quotes.

To remedy this, we explored options such as replacing the single quotes with double quotes. However this led to issues where the name of a game included an apostrophe, e.g. `Assassin's Creed`. 

A workable solution was to read the data carefully using the `ast.literal_eval()` method and saving the outputs as `.json` files.

## User-items data

The dataset was downloaded from https://cseweb.ucsd.edu/~jmcauley/datasets.html#steam_data file *Version 1: User and Item Data*.

# Import libraries

import pandas as pd
import numpy as np
import ast
import json

# Read data using readlines()

with open('australian_users_items.json') as f:
    lines = f.readlines()[:30000]
len(lines)

There are 88310 lines of data in this file, each representing a user.

# Evaluate the first line
# j = ast.literal_eval(lines[0])
# j

# Create a string representing a list with each line seperated by comma
# newstring = "[" + ",".join(lines) + "]"

# Evaluate this new string
j = ast.literal_eval("[" + ",".join(lines) + "]")

# Save file as JSON

with open('data.json', 'w') as json_file:
    json.dump(j, json_file)

We now have a `.json` file that we can easily view as a Pandas DataFrame.

# Load dataframe to check
df = pd.DataFrame(j)
df.head()

## Items detail data

This dataset was downloaded from https://cseweb.ucsd.edu/~jmcauley/datasets.html#steam_data file *Version 2: item metadata*

Similar to above, we will use `ast.literal_eval()` to read the data and then save as a `.json` file.

# Read data

with open('steam_games.json') as f:
    lines = f.readlines()

# View first line

lines[0]

# Get number of lines
len(lines)

There are 32135 lines, each representing a different game.

# evaluate the first string
j = ast.literal_eval(lines[0])
j

# Create a string representing a list with each line seperated by comma
newstring = '[' + ','.join(lines) + ']'

# Evaluate this new string
j = ast.literal_eval(newstring)

# Save file as JSON

with open('gamesdata.json', 'w') as json_file:
    json.dump(j, json_file)

We now have a `.json` file that we can easily view as a Pandas DataFrame.

# Load dataframe to check
df = pd.DataFrame(j)
df.head()
