==> 2-Data-Preprocessing.ipynb <==
# Data Preprocessing

In this Notebook, we will conduct initial preprocessing of the two JSON files we obtained in the previous Notebook. We will aim to create relevant Pandas DataFrames and save files as `csv` for subsequent exploration and modelling. In particular, we will extract user ids and game ids to create a user-item interactions DataFrame, with each row being a particular user-item relationship (namely owned item).

# Import libraries
import pandas as pd
import numpy as np

# Import Warnings
import warnings

warnings.filterwarnings("ignore")


## Games data

We will first load the `gamesdata` file, which has a row for each game and various descriptive features as columns.

# Load games data
df = pd.read_json("gamesdata.json")
df.sample(5)


We note that there are certain features which are lists, namely genres, tags and specs. These will be investigated further in the Data Exploration phase.

# Save as csv
df.to_csv("gamesdata.csv")


## User items Data

We now load the user items data, which has users as rows and details regarding items owned as columns.

# Load users/items data
useritems = pd.read_json("data.json")
useritems.head()


Whilst the `items_count` feature will not be relevant for the user-item interactions/ recommendation engine, it will be relevant to our stakeholders for the purpose of understanding their user base. As such let us extract it and save it for the data exploration phase.

# Extract items_count feat
numgames = useritems[["user_id", "items_count"]]
numgames.head()


# Save as csv
numgames.to_csv("numgames.csv")


We note that the `items` column appears to be a list of dictionaries. Let us look at it in further detail.

# # Preview items column values for first user
# # Restrict to first 2 items in dictionary

# useritems["items"][0][0:2]


Each game is represented by a dictionary with keys the game's `item_id`, `item_name`, `playtime_forever` and `playtime_2weeks`. The dictionaries are then storred in a list.

We will look to extract the `item_id`'s into a seperate column. For now we will leave the playtime data but look to incorporate it later.

# Get all item_id's for first user
gameids = [
    useritems["items"][0][index]["item_id"]
    for index, _ in enumerate(useritems["items"][0])
]
# Show first 10 item ids
gameids[:10]


Extract `item_id` from the dictionaries for each user.

# Create column with item ids
useritems["item_id"] = useritems["items"].apply(
    lambda x: [x[index]["item_id"] for index, _ in enumerate(x)]
)
useritems[["user_id", "items_count", "items", "item_id"]].head(10)
# useritems.head(10)


As the unique user `steam_id` is a large integer, we will replace it with a new `uid` counter, which starts at 0 and increments by 1 (like the index).

We will also only select the relevant columns for the purpose of building a user-item interactions matrix, namely the newly created user id `uid` and the `item_id`.

# Add a column with substitute user_id, counter
useritems["uid"] = np.arange(len(useritems))

# Take relevant columns
useritems = useritems[["uid", "item_id"]]

# Check
useritems.head()


The next step is to explode the `item_id` into seperate rows, so each user-item interaction has it's own row.

# Explode item_ids into seperate rows
lst_col = "item_id"
useritems = pd.DataFrame(
    {
        col: np.repeat(useritems[col].values, useritems[lst_col].str.len())
        for col in useritems.columns.difference([lst_col])
    }
).assign(**{lst_col: np.concatenate(useritems[lst_col].values)})[
    useritems.columns.tolist()
]
useritems


Add a binary column `owned` which will have 1s everywhere, as only items owned appear in the DataFrame.

# Add binary owned column
useritems["owned"] = np.ones(shape=useritems.shape[0])

# Check
useritems.head()


len(useritems)


We note that we have individual user-item relationships represented in our DataFrame.

We want to restrict ourselves to user-item relationships where the item is in the first `gamesdata` DataFrame to be able to extract relevant information such as genre. 

We will ensure that the DataFrames can be merged on the game `id` feature by changing the type and column name.

# Change item_id to int
useritems["item_id"] = useritems["item_id"].astype(int)

# Rename column to match
useritems = useritems.rename(columns={"item_id": "id"})


We can now merge the DataFrames.

# Merge useritems and games data dataframes
alldata = pd.merge(useritems, df, on="id")
alldata.head()
    
alldata[["uid", "title", "genres", "release_date", "tags", "price", "metascore"]].head()

len(alldata)


# Drop entries with no title
datawithnames = alldata.dropna(axis=0, subset=["title"])
datawithnames[
    ["uid", "genres", "title", "release_date", "tags", "specs", "price", "sentiment", "metascore"]
].sample(5)

# datawithnames.columns


len(datawithnames)


We will save this DataFrame as a `csv` file to conduct data exploration and gain statistics.

# Save to csv
datawithnames.to_csv("mergeddata.csv")


Finally, let us extract the relevant columns for our user-item interactions matrix.

# Get relevant columns for recommendation engine
recdata = datawithnames[["uid", "id", "owned"]]
recdata.head()


# Save to csv
recdata.to_csv("recdata.csv")
